Implemented a future weather core logic

+++ app/src/main/java/com/weather/weathermvvmapp/data/database/future_db/FutureWeatherEntity.kt
@@ -21,7 +21,7 @@ data class FutureWeather(
     @TypeConverters(FutureWeatherConverter::class)
     val list: ArrayList<FutureWeatherList>?,
     @SerializedName("message")
-    val message: Int
+    val message: Double
 ) {
     @PrimaryKey(autoGenerate = false)
     var id: Int = FUTURE_WEATHER_ID
